test(cli): split daemon OpenClaw test suite
