+++ packages/cli/test/daemon-openclaw.part-01.test.ts
@@ -1,0 +1,474 @@
+import { EventEmitter } from 'node:events';
+import { existsSync, mkdirSync, mkdtempSync, readFileSync, readdirSync, rmSync, writeFileSync } from 'node:fs';
+import { homedir, tmpdir } from 'node:os';
+import { join } from 'node:path';
+import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
+import {
+  buildOpenClawChannelHeaders,
+  cancelPendingLocalAgentAttachJob,
+  connectLocalAgentIntegrationFromUi,
+  connectLocalAgentIntegration,
+  daemonState,
+  getLocalAgentIntegration,
+  getOpenClawChannelTargets,
+  hasConfiguredLocalAgentChat,
+  hasOpenClawChatTurnContent,
+  isLoopbackClientIp,
+  isOpenClawMemorySlotElected,
+  normalizeOpenClawAttachmentRefs,
+  normalizeOpenClawChatContextEntry,
+  normalizeOpenClawChatContextEntries,
+  isValidOpenClawPersistTurnPayload,
+  listLocalAgentIntegrations,
+  OPENCLAW_CHANNEL_RESPONSE_TIMEOUT_MS,
+  parseRequiredSignatures,
+  pipeOpenClawStream,
+  probeOpenClawChannelHealth,
+  refreshLocalAgentIntegrationFromUi,
+  reverseLocalAgentSetupForUi,
+  runOpenClawUiSetup,
+  verifyOpenClawAttachmentRefsProvenance,
+  normalizeExplicitLocalAgentDisconnectBody,
+  shouldBypassRateLimitForLoopbackTraffic,
+  updateLocalAgentIntegration,
+} from '../src/daemon.js';
+import { handleOpenclawRoutes } from '../src/daemon/routes/openclaw.js';
+import { mergeOpenClawConfig, type AdapterEntryConfig } from '@origintrail-official/dkg-adapter-openclaw';
+import type { DkgConfig } from '../src/config.js';
+
+// Default entryConfig fixture matching the shape `runSetup` builds at
+// Step 5 — same values setup writes into plugins.entries.adapter-openclaw.config.
+const testEntryConfig: AdapterEntryConfig = {
+  daemonUrl: 'http://127.0.0.1:9200',
+  memory: { enabled: true },
+  channel: { enabled: true },
+};
+
+function makeConfig(overrides: Partial<DkgConfig> = {}): DkgConfig {
+  return {
+    name: 'test-node',
+    apiPort: 9200,
+    listenPort: 0,
+    nodeRole: 'edge',
+    ...overrides,
+  };
+}
+
+function makeJsonRequest(method: string, path: string, payload: unknown) {
+  const req = new EventEmitter() as any;
+  req.method = method;
+  req.url = path;
+  req.headers = {};
+  setTimeout(() => {
+    req.emit('data', Buffer.from(JSON.stringify(payload)));
+    req.emit('end');
+  }, 0);
+  return req;
+}
+
+function makeJsonResponse() {
+  const res = new EventEmitter() as any;
+  res.statusCode = 0;
+  res.headers = {};
+  res.body = '';
+  res.writableEnded = false;
+  res.writeHead = (status: number, headers: Record<string, string>) => {
+    res.statusCode = status;
+    res.headers = headers;
+  };
+  res.write = (chunk: string | Buffer) => {
+    res.body += Buffer.isBuffer(chunk) ? chunk.toString('utf8') : chunk;
+    return true;
+  };
+  res.end = (chunk?: string | Buffer) => {
+    if (chunk) res.write(chunk);
+    res.writableEnded = true;
+  };
+  return res;
+}
+
+function freshExtractionStatusTimes() {
+  const completedAt = new Date().toISOString();
+  const startedAt = new Date(Date.now() - 1000).toISOString();
+  return { startedAt, completedAt };
+}
+
+function makeOpenClawRouteContext(
+  payload: unknown,
+  path = '/api/openclaw-channel/send',
+  configOverrides: Partial<DkgConfig> = {},
+) {
+  const req = makeJsonRequest('POST', path, payload);
+  const res = makeJsonResponse();
+  return {
+    ctx: {
+      req,
+      res,
+      agent: {
+        store: { query: vi.fn(async () => ({ bindings: [] })) },
+      },
+      config: makeConfig({
+        localAgentIntegrations: {
+          openclaw: {
+            enabled: true,
+            capabilities: { localChat: true },
+            transport: { kind: 'openclaw-channel', bridgeUrl: 'http://127.0.0.1:9301' },
+          },
+        },
+        ...configOverrides,
+      }),
+      memoryManager: {
+        storeChatExchange: vi.fn(async () => {}),
+      },
+      bridgeAuthToken: 'bridge-token',
+      extractionStatus: new Map(),
+      path,
+      requestAgentAddress: '0x0000000000000000000000000000000000000001',
+    } as any,
+    res,
+  };
+}
+
+
+
+describe('OpenClaw channel routing helpers', () => {
+  beforeEach(() => {
+    daemonState.openClawBridgeHealth = null;
+  });
+
+
+
+  it('keeps the public context normalizer backward-compatible with legacy attachment import entries', () => {
+    const contextEntry = {
+      key: 'target_context_graph',
+      label: 'Target context graph',
+      value: 'Project One (cg1)',
+    };
+    const legacyEntry = {
+      key: 'attachment_import_result_legacy',
+      label: 'Attachment import result: skipped.epub',
+      value: [
+        'fileName=skipped.epub',
+        'assertionUri=did:dkg:context-graph:cg1/assertion/skipped',
+        'contextGraphId=cg1',
+        'fileHash=sha256:skip',
+        'contentType=application/epub+zip',
+        'extractionStatus=skipped',
+        'pipelineUsed=none',
+        'tripleCount=0',
+      ].join('; '),
+    };
+
+    expect(normalizeOpenClawChatContextEntry(legacyEntry)).toEqual(legacyEntry);
+    expect(normalizeOpenClawChatContextEntries(undefined)).toBeUndefined();
+    expect(normalizeOpenClawChatContextEntries(null)).toBeUndefined();
+    expect(normalizeOpenClawChatContextEntries([
+      contextEntry,
+      legacyEntry,
+    ])).toEqual([contextEntry, legacyEntry]);
+  });
+
+  it('defaults to the standalone bridge when no gateway transport is configured', () => {
+    expect(getOpenClawChannelTargets(makeConfig())).toEqual([
+      {
+        name: 'bridge',
+        inboundUrl: 'http://127.0.0.1:9201/inbound',
+        streamUrl: 'http://127.0.0.1:9201/inbound/stream',
+        healthUrl: 'http://127.0.0.1:9201/health',
+      },
+    ]);
+  });
+
+  it('uses only the gateway route when gatewayUrl is configured without an explicit bridgeUrl', () => {
+    expect(getOpenClawChannelTargets(makeConfig({
+      localAgentIntegrations: {
+        openclaw: {
+          enabled: true,
+          transport: {
+            kind: 'openclaw-channel',
+            gatewayUrl: 'http://gateway.local:3030',
+          },
+        },
+      },
+    }))).toEqual([
+      {
+        name: 'gateway',
+        inboundUrl: 'http://gateway.local:3030/api/dkg-channel/inbound',
+        healthUrl: 'http://gateway.local:3030/api/dkg-channel/health',
+      },
+    ]);
+  });
+
+  it('keeps both transports when bridgeUrl and gatewayUrl are both configured', () => {
+    expect(getOpenClawChannelTargets(makeConfig({
+      localAgentIntegrations: {
+        openclaw: {
+          enabled: true,
+          transport: {
+            kind: 'openclaw-channel',
+            bridgeUrl: 'http://127.0.0.1:9301',
+            gatewayUrl: 'http://gateway.local:3030',
+          },
+        },
+      },
+    }))).toEqual([
+      {
+        name: 'bridge',
+        inboundUrl: 'http://127.0.0.1:9301/inbound',
+        streamUrl: 'http://127.0.0.1:9301/inbound/stream',
+        healthUrl: 'http://127.0.0.1:9301/health',
+      },
+      {
+        name: 'gateway',
+        inboundUrl: 'http://gateway.local:3030/api/dkg-channel/inbound',
+        healthUrl: 'http://gateway.local:3030/api/dkg-channel/health',
+      },
+    ]);
+  });
+
+  it('uses an explicit bridge health URL without rewriting the gateway health target', () => {
+    expect(getOpenClawChannelTargets(makeConfig({
+      localAgentIntegrations: {
+        openclaw: {
+          enabled: true,
+          transport: {
+            kind: 'openclaw-channel',
+            bridgeUrl: 'http://127.0.0.1:9301',
+            gatewayUrl: 'http://gateway.local:3030',
+            healthUrl: 'http://127.0.0.1:9301/custom-health',
+          },
+        },
+      },
+    }))).toEqual([
+      {
+        name: 'bridge',
+        inboundUrl: 'http://127.0.0.1:9301/inbound',
+        streamUrl: 'http://127.0.0.1:9301/inbound/stream',
+        healthUrl: 'http://127.0.0.1:9301/custom-health',
+      },
+      {
+        name: 'gateway',
+        inboundUrl: 'http://gateway.local:3030/api/dkg-channel/inbound',
+        healthUrl: 'http://gateway.local:3030/api/dkg-channel/health',
+      },
+    ]);
+  });
+
+  it('leaves an ambiguous explicit health URL unclassified when both bridge and gateway exist', () => {
+    expect(getOpenClawChannelTargets(makeConfig({
+      localAgentIntegrations: {
+        openclaw: {
+          enabled: true,
+          transport: {
+            kind: 'openclaw-channel',
+            bridgeUrl: 'http://bridge.local:9301',
+            gatewayUrl: 'http://gateway.local:3030',
+            healthUrl: 'http://bridge-health.local/custom-health',
+          },
+        },
+      },
+    }))).toEqual([
+      {
+        name: 'bridge',
+        inboundUrl: 'http://bridge.local:9301/inbound',
+        streamUrl: 'http://bridge.local:9301/inbound/stream',
+        healthUrl: 'http://bridge.local:9301/health',
+      },
+      {
+        name: 'gateway',
+        inboundUrl: 'http://gateway.local:3030/api/dkg-channel/inbound',
+        healthUrl: 'http://gateway.local:3030/api/dkg-channel/health',
+      },
+    ]);
+  });
+
+  it('uses an explicit gateway health URL without rewriting the bridge health target', () => {
+    expect(getOpenClawChannelTargets(makeConfig({
+      localAgentIntegrations: {
+        openclaw: {
+          enabled: true,
+          transport: {
+            kind: 'openclaw-channel',
+            bridgeUrl: 'http://127.0.0.1:9301',
+            gatewayUrl: 'http://gateway.local:3030',
+            healthUrl: 'http://gateway.local:3030/api/dkg-channel/health/probe',
+          },
+        },
+      },
+    }))).toEqual([
+      {
+        name: 'bridge',
+        inboundUrl: 'http://127.0.0.1:9301/inbound',
+        streamUrl: 'http://127.0.0.1:9301/inbound/stream',
+        healthUrl: 'http://127.0.0.1:9301/health',
+      },
+      {
+        name: 'gateway',
+        inboundUrl: 'http://gateway.local:3030/api/dkg-channel/inbound',
+        healthUrl: 'http://gateway.local:3030/api/dkg-channel/health/probe',
+      },
+    ]);
+  });
+
+  it('uses explicit gateway health by port when bridge and gateway share a host', () => {
+    expect(getOpenClawChannelTargets(makeConfig({
+      localAgentIntegrations: {
+        openclaw: {
+          enabled: true,
+          transport: {
+            kind: 'openclaw-channel',
+            bridgeUrl: 'http://localhost:9301',
+            gatewayUrl: 'http://localhost:3030',
+            healthUrl: 'http://localhost:3030/api/dkg-channel/health/probe',
+          },
+        },
+      },
+    }))).toEqual([
+      {
+        name: 'bridge',
+        inboundUrl: 'http://localhost:9301/inbound',
+        streamUrl: 'http://localhost:9301/inbound/stream',
+        healthUrl: 'http://localhost:9301/health',
+      },
+      {
+        name: 'gateway',
+        inboundUrl: 'http://localhost:3030/api/dkg-channel/inbound',
+        healthUrl: 'http://localhost:3030/api/dkg-channel/health/probe',
+      },
+    ]);
+  });
+
+  it('prefers a nested gateway health URL over a broader bridge base', () => {
+    expect(getOpenClawChannelTargets(makeConfig({
+      localAgentIntegrations: {
+        openclaw: {
+          enabled: true,
+          transport: {
+            kind: 'openclaw-channel',
+            bridgeUrl: 'http://localhost:9301',
+            gatewayUrl: 'http://localhost:9301/api/dkg-channel',
+            healthUrl: 'http://localhost:9301/api/dkg-channel/health',
+          },
+        },
+      },
+    }))).toEqual([
+      {
+        name: 'bridge',
+        inboundUrl: 'http://localhost:9301/inbound',
+        streamUrl: 'http://localhost:9301/inbound/stream',
+        healthUrl: 'http://localhost:9301/health',
+      },
+      {
+        name: 'gateway',
+        inboundUrl: 'http://localhost:9301/api/dkg-channel/inbound',
+        healthUrl: 'http://localhost:9301/api/dkg-channel/health',
+      },
+    ]);
+  });
+
+  it('uses an explicit gateway health URL when only gateway transport is configured', () => {
+    expect(getOpenClawChannelTargets(makeConfig({
+      localAgentIntegrations: {
+        openclaw: {
+          enabled: true,
+          transport: {
+            kind: 'openclaw-channel',
+            gatewayUrl: 'http://gateway.local:3030',
+            healthUrl: 'http://gateway.local:3030/api/dkg-channel/health/probe',
+          },
+        },
+      },
+    }))).toEqual([
+      {
+        name: 'gateway',
+        inboundUrl: 'http://gateway.local:3030/api/dkg-channel/inbound',
+        healthUrl: 'http://gateway.local:3030/api/dkg-channel/health/probe',
+      },
+    ]);
+  });
+
+  it('ignores an explicit bridge health URL outside the bridge base', () => {
+    expect(getOpenClawChannelTargets(makeConfig({
+      localAgentIntegrations: {
+        openclaw: {
+          enabled: true,
+          transport: {
+            kind: 'openclaw-channel',
+            bridgeUrl: 'http://127.0.0.1:9301',
+            healthUrl: 'http://bridge-health.local/custom-health',
+          },
+        },
+      },
+    }))).toEqual([
+      {
+        name: 'bridge',
+        inboundUrl: 'http://127.0.0.1:9301/inbound',
+        streamUrl: 'http://127.0.0.1:9301/inbound/stream',
+        healthUrl: 'http://127.0.0.1:9301/health',
+      },
+    ]);
+  });
+
+  it('prefers the generic local agent integration transport when OpenClaw is connected through the registry', () => {
+    expect(getOpenClawChannelTargets(makeConfig({
+      localAgentIntegrations: {
+        openclaw: {
+          enabled: true,
+          transport: {
+            kind: 'openclaw-channel',
+            bridgeUrl: 'http://127.0.0.1:9401',
+            gatewayUrl: 'http://gateway.local:4040',
+          },
+        },
+      },
+    }))).toEqual([
+      {
+        name: 'bridge',
+        inboundUrl: 'http://127.0.0.1:9401/inbound',
+        streamUrl: 'http://127.0.0.1:9401/inbound/stream',
+        healthUrl: 'http://127.0.0.1:9401/health',
+      },
+      {
+        name: 'gateway',
+        inboundUrl: 'http://gateway.local:4040/api/dkg-channel/inbound',
+        healthUrl: 'http://gateway.local:4040/api/dkg-channel/health',
+      },
+    ]);
+  });
+
+  it('returns no OpenClaw channel targets when the registry explicitly disables the integration', () => {
+    expect(getOpenClawChannelTargets(makeConfig({
+      localAgentIntegrations: {
+        openclaw: {
+          enabled: false,
+        },
+      },
+    }))).toEqual([]);
+  });
+
+  it('adds the bridge auth header only for standalone bridge requests', () => {
+    const bridgeHeaders = buildOpenClawChannelHeaders(
+      {
+        name: 'bridge',
+        inboundUrl: 'http://127.0.0.1:9201/inbound',
+      },
+      'secret-token',
+      { 'Content-Type': 'application/json' },
+    );
+    expect(bridgeHeaders).toEqual({
+      'Content-Type': 'application/json',
+      'x-dkg-bridge-token': 'secret-token',
+    });
+
+    const gatewayHeaders = buildOpenClawChannelHeaders(
+      {
+        name: 'gateway',
+        inboundUrl: 'http://gateway.local/api/dkg-channel/inbound',
+      },
+      'secret-token',
+      { 'Content-Type': 'application/json' },
+    );
+    expect(gatewayHeaders).toEqual({ 'Content-Type': 'application/json' });
+  });
+
+});
