+++ packages/cli/test/daemon-openclaw.part-09.test.ts
@@ -1,0 +1,274 @@
+import { EventEmitter } from 'node:events';
+import { existsSync, mkdirSync, mkdtempSync, readFileSync, readdirSync, rmSync, writeFileSync } from 'node:fs';
+import { homedir, tmpdir } from 'node:os';
+import { join } from 'node:path';
+import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
+import {
+  buildOpenClawChannelHeaders,
+  cancelPendingLocalAgentAttachJob,
+  connectLocalAgentIntegrationFromUi,
+  connectLocalAgentIntegration,
+  daemonState,
+  getLocalAgentIntegration,
+  getOpenClawChannelTargets,
+  hasConfiguredLocalAgentChat,
+  hasOpenClawChatTurnContent,
+  isLoopbackClientIp,
+  isOpenClawMemorySlotElected,
+  normalizeOpenClawAttachmentRefs,
+  normalizeOpenClawChatContextEntry,
+  normalizeOpenClawChatContextEntries,
+  isValidOpenClawPersistTurnPayload,
+  listLocalAgentIntegrations,
+  OPENCLAW_CHANNEL_RESPONSE_TIMEOUT_MS,
+  parseRequiredSignatures,
+  pipeOpenClawStream,
+  probeOpenClawChannelHealth,
+  refreshLocalAgentIntegrationFromUi,
+  reverseLocalAgentSetupForUi,
+  runOpenClawUiSetup,
+  verifyOpenClawAttachmentRefsProvenance,
+  normalizeExplicitLocalAgentDisconnectBody,
+  shouldBypassRateLimitForLoopbackTraffic,
+  updateLocalAgentIntegration,
+} from '../src/daemon.js';
+import { handleOpenclawRoutes } from '../src/daemon/routes/openclaw.js';
+import { mergeOpenClawConfig, type AdapterEntryConfig } from '@origintrail-official/dkg-adapter-openclaw';
+import type { DkgConfig } from '../src/config.js';
+
+// Default entryConfig fixture matching the shape `runSetup` builds at
+// Step 5 — same values setup writes into plugins.entries.adapter-openclaw.config.
+const testEntryConfig: AdapterEntryConfig = {
+  daemonUrl: 'http://127.0.0.1:9200',
+  memory: { enabled: true },
+  channel: { enabled: true },
+};
+
+function makeConfig(overrides: Partial<DkgConfig> = {}): DkgConfig {
+  return {
+    name: 'test-node',
+    apiPort: 9200,
+    listenPort: 0,
+    nodeRole: 'edge',
+    ...overrides,
+  };
+}
+
+function makeJsonRequest(method: string, path: string, payload: unknown) {
+  const req = new EventEmitter() as any;
+  req.method = method;
+  req.url = path;
+  req.headers = {};
+  setTimeout(() => {
+    req.emit('data', Buffer.from(JSON.stringify(payload)));
+    req.emit('end');
+  }, 0);
+  return req;
+}
+
+function makeJsonResponse() {
+  const res = new EventEmitter() as any;
+  res.statusCode = 0;
+  res.headers = {};
+  res.body = '';
+  res.writableEnded = false;
+  res.writeHead = (status: number, headers: Record<string, string>) => {
+    res.statusCode = status;
+    res.headers = headers;
+  };
+  res.write = (chunk: string | Buffer) => {
+    res.body += Buffer.isBuffer(chunk) ? chunk.toString('utf8') : chunk;
+    return true;
+  };
+  res.end = (chunk?: string | Buffer) => {
+    if (chunk) res.write(chunk);
+    res.writableEnded = true;
+  };
+  return res;
+}
+
+function freshExtractionStatusTimes() {
+  const completedAt = new Date().toISOString();
+  const startedAt = new Date(Date.now() - 1000).toISOString();
+  return { startedAt, completedAt };
+}
+
+function makeOpenClawRouteContext(
+  payload: unknown,
+  path = '/api/openclaw-channel/send',
+  configOverrides: Partial<DkgConfig> = {},
+) {
+  const req = makeJsonRequest('POST', path, payload);
+  const res = makeJsonResponse();
+  return {
+    ctx: {
+      req,
+      res,
+      agent: {
+        store: { query: vi.fn(async () => ({ bindings: [] })) },
+      },
+      config: makeConfig({
+        localAgentIntegrations: {
+          openclaw: {
+            enabled: true,
+            capabilities: { localChat: true },
+            transport: { kind: 'openclaw-channel', bridgeUrl: 'http://127.0.0.1:9301' },
+          },
+        },
+        ...configOverrides,
+      }),
+      memoryManager: {
+        storeChatExchange: vi.fn(async () => {}),
+      },
+      bridgeAuthToken: 'bridge-token',
+      extractionStatus: new Map(),
+      path,
+      requestAgentAddress: '0x0000000000000000000000000000000000000001',
+    } as any,
+    res,
+  };
+}
+
+describe('OpenClaw persist-turn validation', () => {
+
+
+
+  it('rejects completed attachment refs backed only by skipped import metadata', async () => {
+    const attachmentRefs = [{
+      assertionUri: 'did:dkg:context-graph:cg1/assertion/chat-doc',
+      fileHash: 'sha256:abc123',
+      contextGraphId: 'cg1',
+      fileName: 'chat-doc.pdf',
+      extractionStatus: 'completed' as const,
+      tripleCount: 0,
+    }];
+    const store = {
+      query: async () => ({
+        bindings: [{
+          fileHash: '"sha256:abc123"',
+          contentType: '"application/pdf"',
+          extractionStatus: '"skipped"',
+          tripleCount: '"0"^^<http://www.w3.org/2001/XMLSchema#integer>',
+          sourceFileName: '"chat-doc.pdf"',
+        }],
+      }),
+    };
+
+    await expect(
+      verifyOpenClawAttachmentRefsProvenance({ store } as any, new Map(), attachmentRefs),
+    ).resolves.toBeUndefined();
+  });
+
+  it('rejects completed attachment refs after the extraction cache entry is gone and the meta graph no longer has the assertion', async () => {
+    const attachmentRefs = [{
+      assertionUri: 'did:dkg:context-graph:cg1/assertion/chat-doc',
+      fileHash: 'sha256:abc123',
+      contextGraphId: 'cg1',
+      fileName: 'chat-doc.pdf',
+      extractionStatus: 'completed' as const,
+    }];
+    const queryCalls: unknown[][] = [];
+    const store = {
+      query: async (...args: unknown[]) => { queryCalls.push(args); return { bindings: [] }; },
+    };
+
+    await expect(
+      verifyOpenClawAttachmentRefsProvenance({ store } as any, new Map(), attachmentRefs),
+    ).resolves.toBeUndefined();
+    expect(queryCalls).toHaveLength(1);
+  });
+
+  it('rejects attachment refs when the stored source file name does not match', async () => {
+    const attachmentRefs = [{
+      assertionUri: 'did:dkg:context-graph:cg1/assertion/chat-doc',
+      fileHash: 'sha256:abc123',
+      contextGraphId: 'cg1',
+      fileName: 'spoofed.pdf',
+      extractionStatus: 'completed' as const,
+    }];
+    const store = {
+      query: async () => ({
+        bindings: [{
+          fileHash: '"sha256:abc123"',
+          sourceFileName: '"chat-doc.pdf"',
+        }],
+      }),
+    };
+
+    await expect(
+      verifyOpenClawAttachmentRefsProvenance({ store } as any, new Map(), attachmentRefs),
+    ).resolves.toBeUndefined();
+  });
+
+  it('rejects forged attachment refs when graph metadata does not match', async () => {
+    const attachmentRefs = [{
+      assertionUri: 'did:dkg:context-graph:cg1/assertion/chat-doc',
+      fileHash: 'sha256:forged',
+      contextGraphId: 'cg1',
+      fileName: 'chat-doc.pdf',
+      extractionStatus: 'completed' as const,
+    }];
+    const store = {
+      query: async () => ({
+        bindings: [{
+          fileHash: '"sha256:real"',
+          contentType: '"application/pdf"',
+          tripleCount: '"42"^^<http://www.w3.org/2001/XMLSchema#integer>',
+        }],
+      }),
+    };
+
+    await expect(
+      verifyOpenClawAttachmentRefsProvenance({ store } as any, new Map(), attachmentRefs),
+    ).resolves.toBeUndefined();
+  });
+
+  it('accepts explicit failed and pending persistence states', () => {
+    expect(isValidOpenClawPersistTurnPayload({
+      sessionId: 'openclaw:dkg-ui',
+      userMessage: 'hi',
+      assistantReply: '',
+      persistenceState: 'failed',
+    })).toBe(true);
+    expect(isValidOpenClawPersistTurnPayload({
+      sessionId: 'openclaw:dkg-ui',
+      userMessage: 'hi',
+      assistantReply: '',
+      persistenceState: 'pending',
+    })).toBe(true);
+  });
+
+  it('rejects unknown persistence states', () => {
+    expect(isValidOpenClawPersistTurnPayload({
+      sessionId: 'openclaw:dkg-ui',
+      userMessage: 'hi',
+      assistantReply: '',
+      persistenceState: 'cancelled',
+    })).toBe(false);
+  });
+
+  it('accepts string/null failure reasons and rejects invalid ones', () => {
+    expect(isValidOpenClawPersistTurnPayload({
+      sessionId: 'openclaw:dkg-ui',
+      userMessage: 'hi',
+      assistantReply: '',
+      persistenceState: 'failed',
+      failureReason: 'timeout',
+    })).toBe(true);
+    expect(isValidOpenClawPersistTurnPayload({
+      sessionId: 'openclaw:dkg-ui',
+      userMessage: 'hi',
+      assistantReply: '',
+      persistenceState: 'failed',
+      failureReason: null,
+    })).toBe(true);
+    expect(isValidOpenClawPersistTurnPayload({
+      sessionId: 'openclaw:dkg-ui',
+      userMessage: 'hi',
+      assistantReply: '',
+      persistenceState: 'failed',
+      failureReason: { code: 'timeout' },
+    })).toBe(false);
+  });
+
+});
