+++ packages/cli/test/daemon-openclaw.part-13.test.ts
@@ -1,0 +1,246 @@
+import { EventEmitter } from 'node:events';
+import { existsSync, mkdirSync, mkdtempSync, readFileSync, readdirSync, rmSync, writeFileSync } from 'node:fs';
+import { homedir, tmpdir } from 'node:os';
+import { join } from 'node:path';
+import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
+import {
+  buildOpenClawChannelHeaders,
+  cancelPendingLocalAgentAttachJob,
+  connectLocalAgentIntegrationFromUi,
+  connectLocalAgentIntegration,
+  daemonState,
+  getLocalAgentIntegration,
+  getOpenClawChannelTargets,
+  hasConfiguredLocalAgentChat,
+  hasOpenClawChatTurnContent,
+  isLoopbackClientIp,
+  isOpenClawMemorySlotElected,
+  normalizeOpenClawAttachmentRefs,
+  normalizeOpenClawChatContextEntry,
+  normalizeOpenClawChatContextEntries,
+  isValidOpenClawPersistTurnPayload,
+  listLocalAgentIntegrations,
+  OPENCLAW_CHANNEL_RESPONSE_TIMEOUT_MS,
+  parseRequiredSignatures,
+  pipeOpenClawStream,
+  probeOpenClawChannelHealth,
+  refreshLocalAgentIntegrationFromUi,
+  reverseLocalAgentSetupForUi,
+  runOpenClawUiSetup,
+  verifyOpenClawAttachmentRefsProvenance,
+  normalizeExplicitLocalAgentDisconnectBody,
+  shouldBypassRateLimitForLoopbackTraffic,
+  updateLocalAgentIntegration,
+} from '../src/daemon.js';
+import { handleOpenclawRoutes } from '../src/daemon/routes/openclaw.js';
+import { mergeOpenClawConfig, type AdapterEntryConfig } from '@origintrail-official/dkg-adapter-openclaw';
+import type { DkgConfig } from '../src/config.js';
+
+// Default entryConfig fixture matching the shape `runSetup` builds at
+// Step 5 — same values setup writes into plugins.entries.adapter-openclaw.config.
+const testEntryConfig: AdapterEntryConfig = {
+  daemonUrl: 'http://127.0.0.1:9200',
+  memory: { enabled: true },
+  channel: { enabled: true },
+};
+
+function makeConfig(overrides: Partial<DkgConfig> = {}): DkgConfig {
+  return {
+    name: 'test-node',
+    apiPort: 9200,
+    listenPort: 0,
+    nodeRole: 'edge',
+    ...overrides,
+  };
+}
+
+function makeJsonRequest(method: string, path: string, payload: unknown) {
+  const req = new EventEmitter() as any;
+  req.method = method;
+  req.url = path;
+  req.headers = {};
+  setTimeout(() => {
+    req.emit('data', Buffer.from(JSON.stringify(payload)));
+    req.emit('end');
+  }, 0);
+  return req;
+}
+
+function makeJsonResponse() {
+  const res = new EventEmitter() as any;
+  res.statusCode = 0;
+  res.headers = {};
+  res.body = '';
+  res.writableEnded = false;
+  res.writeHead = (status: number, headers: Record<string, string>) => {
+    res.statusCode = status;
+    res.headers = headers;
+  };
+  res.write = (chunk: string | Buffer) => {
+    res.body += Buffer.isBuffer(chunk) ? chunk.toString('utf8') : chunk;
+    return true;
+  };
+  res.end = (chunk?: string | Buffer) => {
+    if (chunk) res.write(chunk);
+    res.writableEnded = true;
+  };
+  return res;
+}
+
+function freshExtractionStatusTimes() {
+  const completedAt = new Date().toISOString();
+  const startedAt = new Date(Date.now() - 1000).toISOString();
+  return { startedAt, completedAt };
+}
+
+function makeOpenClawRouteContext(
+  payload: unknown,
+  path = '/api/openclaw-channel/send',
+  configOverrides: Partial<DkgConfig> = {},
+) {
+  const req = makeJsonRequest('POST', path, payload);
+  const res = makeJsonResponse();
+  return {
+    ctx: {
+      req,
+      res,
+      agent: {
+        store: { query: vi.fn(async () => ({ bindings: [] })) },
+      },
+      config: makeConfig({
+        localAgentIntegrations: {
+          openclaw: {
+            enabled: true,
+            capabilities: { localChat: true },
+            transport: { kind: 'openclaw-channel', bridgeUrl: 'http://127.0.0.1:9301' },
+          },
+        },
+        ...configOverrides,
+      }),
+      memoryManager: {
+        storeChatExchange: vi.fn(async () => {}),
+      },
+      bridgeAuthToken: 'bridge-token',
+      extractionStatus: new Map(),
+      path,
+      requestAgentAddress: '0x0000000000000000000000000000000000000001',
+    } as any,
+    res,
+  };
+}
+
+describe('local agent integration registry helpers', () => {
+
+
+
+  it('rechecks OpenClaw bridge health quickly after a cached failure so recovery is not sticky', async () => {
+    vi.useFakeTimers();
+    vi.setSystemTime(new Date('2026-04-13T12:00:00.000Z'));
+    const origFetch = globalThis.fetch;
+    let fetchCallCount = 0;
+    const fetchResponses = [
+      new Response(
+        JSON.stringify({ ok: false, error: 'offline' }),
+        { status: 503, headers: { 'Content-Type': 'application/json' } },
+      ),
+      new Response(
+        JSON.stringify({ ok: true, channel: 'dkg-ui' }),
+        { status: 200, headers: { 'Content-Type': 'application/json' } },
+      ),
+    ];
+    globalThis.fetch = (async () => fetchResponses[fetchCallCount++]) as typeof fetch;
+
+    try {
+      const config = makeConfig();
+      const first = await probeOpenClawChannelHealth(config, 'bridge-token', { ignoreBridgeCache: true });
+      expect(first.ok).toBe(false);
+
+      vi.setSystemTime(new Date('2026-04-13T12:00:01.500Z'));
+      const second = await probeOpenClawChannelHealth(config, 'bridge-token');
+
+      expect(second.ok).toBe(true);
+      expect(fetchCallCount).toBe(2);
+    } finally {
+      globalThis.fetch = origFetch;
+      vi.useRealTimers();
+    }
+  });
+
+  it('updates a stored integration without dropping nested metadata', () => {
+    const config = makeConfig({
+      localAgentIntegrations: {
+        hermes: {
+          enabled: true,
+          metadata: {
+            source: 'ui',
+            manifestMode: 'setup-entry',
+          },
+          runtime: {
+            status: 'connecting',
+          },
+        },
+      },
+    });
+
+    const integration = updateLocalAgentIntegration(config, 'hermes', {
+      runtime: {
+        status: 'ready',
+        ready: true,
+      },
+      metadata: {
+        manifestVersion: '1.2.3',
+      },
+    }, new Date('2026-04-13T10:15:00.000Z'));
+
+    expect(integration.status).toBe('ready');
+    expect(integration.runtime.ready).toBe(true);
+    expect(integration.metadata).toEqual({
+      source: 'ui',
+      manifestMode: 'setup-entry',
+      manifestVersion: '1.2.3',
+    });
+  });
+
+  it('replaces stored transport hints so stale OpenClaw gateway URLs do not linger', () => {
+    const config = makeConfig({
+      localAgentIntegrations: {
+        openclaw: {
+          enabled: true,
+          transport: {
+            kind: 'openclaw-channel',
+            gatewayUrl: 'http://gateway.local:3030',
+          },
+          runtime: {
+            status: 'ready',
+            ready: true,
+          },
+        },
+      },
+    }) as DkgConfig & {
+      openclawAdapter?: boolean;
+      openclawChannel?: { gatewayUrl?: string };
+    };
+    config.openclawAdapter = true;
+    config.openclawChannel = {
+      gatewayUrl: 'http://gateway.local:3030',
+    };
+
+    const integration = updateLocalAgentIntegration(config, 'openclaw', {
+      transport: {
+        kind: 'openclaw-channel',
+        bridgeUrl: 'http://127.0.0.1:9301',
+        healthUrl: 'http://127.0.0.1:9301/health',
+      },
+      runtime: {
+        status: 'ready',
+        ready: true,
+      },
+    }, new Date('2026-04-13T10:45:00.000Z'));
+
+    expect(integration.transport.bridgeUrl).toBe('http://127.0.0.1:9301');
+    expect(integration.transport.gatewayUrl).toBeUndefined();
+    expect((config as Record<string, unknown>).openclawAdapter).toBeUndefined();
+    expect((config as Record<string, unknown>).openclawChannel).toBeUndefined();
+  });
+
+});
